Scrapper/requirements.txt
confluent-kafka
selenium
chromedriver-autoinstaller
python-dotenv
faust